feat: implement admin, user, and todo management routers with User login and authentication

+++ fastAPI/FastAPI_Tutorials/05_Auth/ToDo_Project/routers/todos.py
@@ -9,6 +9,7 @@ from starlette import status
 from models import Todos
 from db import SessionLocal
 from Todo import Todo
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -24,20 +25,25 @@ def get_db():
         db.close()
 
 DbDependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict,Depends(get_current_user)]
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: DbDependency):
+async def read_all(user: user_dependency,db: DbDependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
     try:
-        get_data = db.query(Todos).all()
+        get_data = db.query(Todos).filter(user.get("user_id") == Todos.owner).all()
     except SQLAlchemyError as e:
         raise HTTPException(status_code=500, detail=f"Database error occurred {e}")
 
     return get_data
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_by_id(db: DbDependency, todo_id: int = Path(gt=0)):
+async def read_by_id(user: user_dependency,db: DbDependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
     try:
-        todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+        todo_model = db.query(Todos).filter(Todos.owner == user.get("user_id"),Todos.id == todo_id).first()
     except SQLAlchemyError as e:
         raise HTTPException(status_code=500, detail=f"Database error occurred {e}")
 
@@ -46,14 +52,19 @@ async def read_by_id(db: DbDependency, todo_id: int = Path(gt=0)):
     raise HTTPException(status_code=404, detail="Todo not found..")
 
 @router.post("/add_task", status_code=status.HTTP_201_CREATED)
-async def add_todo_task(db: DbDependency, todo_req: Todo):
-    todo_model = Todos(**todo_req.dict())
+async def add_todo_task(user: user_dependency, db: DbDependency, todo_req: Todo):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
+
+    todo_model = Todos(**todo_req.dict(), owner=user.get("user_id"))
     db.add(todo_model)
 
 @router.put("/update_task", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo_task(db: DbDependency, todo_req: Todo, todo_id: int = Query(gt=0)):
+async def update_todo_task(user: user_dependency, db: DbDependency, todo_req: Todo, todo_id: int = Query(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
     try:    
-        todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+        todo_model = db.query(Todos).filter(Todos.owner == user.get("user_id"), Todos.id == todo_id).first()
     except SQLAlchemyError as e:
         raise HTTPException(status_code=500, detail=f"Database error occurred {e}")
 
@@ -67,9 +78,11 @@ async def update_todo_task(db: DbDependency, todo_req: Todo, todo_id: int = Quer
     db.add(todo_model)
 
 @router.delete("/task/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: DbDependency, todo_id: int = Path(gt=0)):
+async def delete_todo(user: user_dependency, db: DbDependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication Failed")
     try:
-        deleted = db.query(Todos).filter(Todos.id == todo_id).delete()
+        deleted = db.query(Todos).filter(Todos.owner == user.get("user_id"), Todos.id == todo_id).delete()
     except SQLAlchemyError as e:
         raise HTTPException(status_code=500, detail=f"Database error occurred {e}")
     if deleted == 0:
